refactor(arch): delete backward-compat shim modules and migrate callers to canonical paths
