--- python/bloqade/lanes/_wrapper.py
@@ -1,9 +1,0 @@
-"""Backward-compatible re-exports of ``RustWrapper`` and ``KirinRustWrapper``.
-
-The canonical import path is ``bloqade.lanes.bytecode._wrapper``.
-"""
-
-from bloqade.lanes.bytecode._wrapper import (
-    KirinRustWrapper as KirinRustWrapper,
-    RustWrapper as RustWrapper,
-)
