+++ python/bloqade/lanes/analysis/atom/atom_state_data.py
@@ -5,9 +5,9 @@ from types import MappingProxyType
 
 from kirin.interp import InterpreterError
 
-from bloqade.lanes._wrapper import RustWrapper
 from bloqade.lanes.arch.spec import ArchSpec
 from bloqade.lanes.bytecode import AtomStateData as _RustAtomStateData
+from bloqade.lanes.bytecode._wrapper import RustWrapper
 from bloqade.lanes.bytecode.encoding import LaneAddress, LocationAddress, ZoneAddress
