--- python/bloqade/lanes/arch/arch_builder.py
@@ -1,14 +1,0 @@
-"""Backward-compatible re-exports of ``ZoneBuilder`` and ``ArchBuilder``.
-
-The canonical import path is ``bloqade.lanes.arch.build.imperative``.
-"""
-
-from bloqade.lanes.arch.build.imperative import (  # noqa: F401
-    ArchBuilder as ArchBuilder,
-    ZoneBuilder as ZoneBuilder,
-    _normalize_index as _normalize_index,
-    _SiteGridQuery as _SiteGridQuery,
-    _to_nm as _to_nm,
-    _validate_aod_rectangle as _validate_aod_rectangle,
-    _WordGridQuery as _WordGridQuery,
-)
