--- python/bloqade/lanes/arch/builder.py
@@ -1,9 +1,0 @@
-"""Backward-compatible re-exports of ``build_arch`` and ``ArchResult``.
-
-The canonical import path is ``bloqade.lanes.arch.build.blueprint``.
-"""
-
-from bloqade.lanes.arch.build.blueprint import (
-    ArchResult as ArchResult,
-    build_arch as build_arch,
-)
