+++ python/bloqade/lanes/arch/generic_full/spec.py
@@ -9,13 +9,18 @@ All zones have full site (hypercube) and word (diagonal) connectivity.
 
 from __future__ import annotations
 
-from bloqade.lanes.arch.builder import ArchResult, build_arch
-from bloqade.lanes.arch.topology import (
+from bloqade.lanes.arch.build.blueprint import (
+    ArchBlueprint,
+    ArchResult,
+    DeviceLayout,
+    ZoneSpec,
+    build_arch,
+)
+from bloqade.lanes.arch.build.topology import (
     DiagonalWordTopology,
     HypercubeSiteTopology,
     MatchingTopology,
 )
-from bloqade.lanes.arch.zone import ArchBlueprint, DeviceLayout, ZoneSpec
 
 _SITES_PER_WORD = 8
 _NUM_ROWS = 8
