--- python/bloqade/lanes/arch/word_factory.py
@@ -1,9 +1,0 @@
-"""Backward-compatible re-exports of ``WordGrid`` and ``create_zone_words``.
-
-The canonical import path is ``bloqade.lanes.arch.build.word_factory``.
-"""
-
-from bloqade.lanes.arch.build.word_factory import (
-    WordGrid as WordGrid,
-    create_zone_words as create_zone_words,
-)
