--- python/bloqade/lanes/arch/zone.py
@@ -1,10 +1,0 @@
-"""Backward-compatible re-exports of ``ArchBlueprint``, ``DeviceLayout``, and ``ZoneSpec``.
-
-The canonical import path is ``bloqade.lanes.arch.build.blueprint``.
-"""
-
-from bloqade.lanes.arch.build.blueprint import (
-    ArchBlueprint as ArchBlueprint,
-    DeviceLayout as DeviceLayout,
-    ZoneSpec as ZoneSpec,
-)
