+++ python/tests/arch/test_gemini_full.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 import pytest
 
+from bloqade.lanes.arch.build.topology import HypercubeSiteTopology
 from bloqade.lanes.arch.gemini_full import get_arch
-from bloqade.lanes.arch.topology import HypercubeSiteTopology
 
 # ── HypercubeSiteTopology: non-power-of-2 extension ──
