+++ python/tests/arch/test_word_factory.py
@@ -1,7 +1,7 @@
 """Tests for word creation helpers."""
 
-from bloqade.lanes.arch.word_factory import WordGrid, create_zone_words
-from bloqade.lanes.arch.zone import DeviceLayout, ZoneSpec
+from bloqade.lanes.arch.build.blueprint import DeviceLayout, ZoneSpec
+from bloqade.lanes.arch.build.word_factory import WordGrid, create_zone_words
 
 
 class TestWordGrid:
